requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
scikit-learn
numpy
openai
supabase
python-jose[cryptography]
